Redraw background properly
Make redrawing of the background when removing or rotating a tile work properly.
Since we are redrawing as little as possible we are drawing at an offset, we must read from the same offset in the background image for this to look right.
The old redrawing read from 0,0 in the background image.